requirements.txt
discord.py==0.16.11
flask==0.12.2
pygsheets==1.1.3
pytz==2017.2
raven==6.3.0
raven-aiohttp==0.6.0
redis==2.10.6
oauth2client==4.1.3